// packages/dashboard/src/components/DashboardGlobalConfigBar/config/enum.ts
export const enum ColorScheme {
  SMART = 'smart'
}

export const enum CardRadius {
  NONE = '0',
  SMALL = '6px',
  BIG = '18px'
}

export const enum CardSpacing {
  COMPACT = 8,
  NORMAL = 16
}

export const enum PagePadding {
  NORMAL = '10px 12px',
  WIDE_LAYOUT = '8px'
}

export const enum LogoPosition {
  TOP = 'TOP',
  LEFT = 'LEFT'
}

export const enum HeaderAlign {
  LEFT = 'felx-start',
  CENTER = 'center'
}